Move change processing functions to a separate module
