+++ crates/corro-agent/src/agent/mod.rs
@@ -12,7 +12,7 @@ mod metrics;
 mod run_root;
 mod setup;
 mod uni;
-mod util;
+pub(crate) mod util;
 
 #[cfg(test)]
 mod tests;
@@ -27,7 +27,6 @@ use uuid::Uuid;
 pub use error::{SyncClientError, SyncRecvError};
 pub use run_root::start_with_config;
 pub use setup::{setup, AgentOptions};
-pub use util::process_multiple_changes;
 
 pub const ANNOUNCE_INTERVAL: Duration = Duration::from_secs(300);
 pub const COMPACT_BOOKED_INTERVAL: Duration = Duration::from_secs(300);
