+++ crates/corro-agent/src/lib.rs
@@ -3,4 +3,5 @@ pub mod agent;
 pub mod api;
 pub mod broadcast;
 pub mod transport;
+pub mod change;
